requirements.txt
fastapi==0.115.1
python-multipart==0.0.9
databases==0.7.0
ormar==0.20.0
aiomysql==0.2.0
mysqlclient==2.2.4
pymysql==1.1.0
pyjwt==2.8.0
redis==5.1.1
uvicorn==0.29.0

celery==5.4.0
cos-python-sdk-v5==1.9.31